Working version: lind-boot besides chain-thread
