+++ src/glibc/stdlib/exit.c
@@ -32,15 +32,16 @@ void __lind_exit(int status) {
 void
 _exit (int status)
 {
-  while (1)
-    {
-      // exit without doing any cleanup
-      __lind_exit(status);
+	//   while (1)
+	//     {
+	//       // exit without doing any cleanup
+	//       __lind_exit(status);
 
-#ifdef ABORT_INSTRUCTION
-      ABORT_INSTRUCTION;
-#endif
-    }
+	// #ifdef ABORT_INSTRUCTION
+	//       ABORT_INSTRUCTION;
+	// #endif
+	//     }
+	  __lind_exit(status);
 }
 rtld_hidden_def (_exit)
 weak_alias (_exit, _Exit)
