+++ src/lind-boot/Cargo.toml
@@ -20,3 +20,4 @@ wasmtime-wasi = { version = "23.0.0", features = ["preview1"] , default-features
 wiggle = { version = "23.0.0", default-features = false }
 anyhow = { version = "1.0.66", default-features = false }
 cap-std = { version = "3.4.2", default-features = false }
+clap = { version = "4", features = ["derive"] }
